% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocationconstraints.R
\name{JAOPuTo_allocationconstraints}
\alias{JAOPuTo_allocationconstraints}
\title{Allocation Constraints}
\usage{
JAOPuTo_allocationconstraints(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the allocation constraints for Core bidding zones
}
\description{
Allocation Constraints
}
\examples{
df <- JAOPuTo_allocationconstraints(as.POSIXct("2022-06-09 00:00", "CET"), as.POSIXct("2022-12-31 23:00", "CET"))
}
