% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congestionincome.R
\name{JAOPuTo_congestionincome}
\alias{JAOPuTo_congestionincome}
\title{Congestion Income}
\usage{
JAOPuTo_congestionincome(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the Congestion Income generated on in Core hubs
}
\description{
Congestion Income
}
\examples{
df <- JAOPuTo_congestionincome(as.POSIXct("2022-06-09 00:00", "CET"), as.POSIXct("2022-12-31 23:00", "CET"))
}
