% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAOPuTo_activeconstraints.R
\name{JAOPuTo_Core_activeconstraints}
\alias{JAOPuTo_Core_activeconstraints}
\title{Core - Active Flow-Based Constraints}
\usage{
JAOPuTo_Core_activeconstraints(start, end)
}
\arguments{
\item{start}{Start datetime; (POSIXct, Date, or character convertible to POSIXct)}

\item{end}{End datetime; (POSIXct, Date, or character convertible to POSIXct)}
}
\value{
A tibble containing the combined data returned by all API requests.
}
\description{
Download active constraints in Core DA FBMC.
}
\examples{
\dontrun{
JAOPuTo_Core_activeconstraints(
  start = "2025-01-01 00:00",
  end = "2025-01-01 01:00"
)
}
}
