% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAOPuTo_Core_activeLTAconstraints.R
\name{JAOPuTo_Core_activeltaconstraints}
\alias{JAOPuTo_Core_activeltaconstraints}
\title{Core - Active LTA Constraints}
\usage{
JAOPuTo_Core_activeltaconstraints(start, end)
}
\arguments{
\item{start}{Start datetime; (POSIXct, Date, or character convertible to POSIXct)}

\item{end}{End datetime; (POSIXct, Date, or character convertible to POSIXct)}
}
\value{
A tibble containing the combined data returned by all API requests.
}
\description{
Download active LTA constraints in Core DA FBMC.
}
\examples{
\dontrun{
JAOPuTo_Core_activeltaconstraints(
  start = "2025-01-01 00:00",
  end = "2025-01-01 01:00"
)
}
}
