% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_allocationconstraints.R
\name{JAOPuTo_Core_allocationconstraints}
\alias{JAOPuTo_Core_allocationconstraints}
\title{Core - Allocation constraints}
\usage{
JAOPuTo_Core_allocationconstraints(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the allocation constraints applicable to selected Core bidding zones' net positions
}
\description{
Core - Allocation constraints
}
\examples{
JAOPuTo_Core_allocationconstraints(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
