% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_alphafactor.R
\name{JAOPuTo_Core_alphafactor}
\alias{JAOPuTo_Core_alphafactor}
\title{Core - Alpha Factor for MCP}
\usage{
JAOPuTo_Core_alphafactor(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the the Alpha Factor from DA Market Clearing Point indicating the share of cross-zonal capacity provided by the FB domain and LTA domain from the SDAC.
}
\description{
Core - Alpha Factor for MCP
}
\examples{
JAOPuTo_Core_alphafactor(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
