% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAOPuTo_Core_atcshadowauctions.R
\name{JAOPuTo_Core_atcshadowauctions}
\alias{JAOPuTo_Core_atcshadowauctions}
\title{Core - Shadow Auction ATC}
\usage{
JAOPuTo_Core_atcshadowauctions(start, end)
}
\arguments{
\item{start}{Start datetime; (POSIXct, Date, or character convertible to POSIXct)}

\item{end}{End datetime; (POSIXct, Date, or character convertible to POSIXct)}
}
\value{
A tibble containing the combined data returned by all API requests.
}
\description{
Download ATCs for shadow auctions in Core DA FBMC.
}
\examples{
\dontrun{
JAOPuTo_Core_atcshadowauctions(
  start = "2025-01-01 00:00",
  end = "2025-01-10 23:00"
)
}
}
