% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAOPuTo_Core_bexrestrictions.R
\name{JAOPuTo_Core_bexrestrictions}
\alias{JAOPuTo_Core_bexrestrictions}
\title{Core - Final Bilateral Exchange Restrictions}
\usage{
JAOPuTo_Core_bexrestrictions(start, end)
}
\arguments{
\item{start}{Start datetime; (POSIXct, Date, or character convertible to POSIXct)}

\item{end}{End datetime; (POSIXct, Date, or character convertible to POSIXct)}
}
\value{
A tibble containing the combined data returned by all API requests.
}
\description{
Download final bilateral exchange restrictions between selected bidding zones in Core DA FBMC.
}
\examples{
\dontrun{
JAOPuTo_Core_bexrestrictions(
  start = "2025-01-01 00:00",
  end = "2025-01-10 23:00"
)
}
}
