% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_congestionincome.R
\name{JAOPuTo_Core_congestionincome}
\alias{JAOPuTo_Core_congestionincome}
\title{Core - Congestion Income}
\usage{
JAOPuTo_Core_congestionincome(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the gross and net congestion income per Core hub, TSO and border
}
\description{
Core - Congestion Income
}
\examples{
JAOPuTo_Core_congestionincome(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
