% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_D2CF.R
\name{JAOPuTo_Core_D2CF}
\alias{JAOPuTo_Core_D2CF}
\title{Core - D2CF}
\usage{
JAOPuTo_Core_D2CF(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the aggregated assumptions from the grid model (D2CF)
}
\description{
Core - D2CF
}
\examples{
JAOPuTo_Core_D2CF(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
