% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_atcexternalborders.R
\name{JAOPuTo_Core_externalATC}
\alias{JAOPuTo_Core_externalATC}
\title{Core - ATC on Core external borders}
\usage{
JAOPuTo_Core_externalATC(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the Available Transfer Capacities (ATC) on Core external borders
}
\description{
Core - ATC on Core external borders
}
\examples{
JAOPuTo_Core_externalATC(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
