% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_finalcomputation.R
\name{JAOPuTo_Core_finalcomputation}
\alias{JAOPuTo_Core_finalcomputation}
\title{Core - Final Computation}
\usage{
JAOPuTo_Core_finalcomputation(StartDateTime, EndDateTime, Presolved = TRUE)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}

\item{Presolved}{Boolean variable (TRUE/FALSE) to determine whether only presolved network elements are included (defaults to TRUE)}
}
\value{
a dataframe containing the final flow-based parameters of the selected business day and MTU following long term nominations (Ltnom balanced).
}
\description{
Core - Final Computation
}
\examples{
JAOPuTo_Core_finalcomputation(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
