% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAOPuTo_CoreID_IDA3_netpositions.R
\name{JAOPuTo_Core_IDA3_netposition}
\alias{JAOPuTo_Core_IDA3_netposition}
\title{Core IDA3 - Net Position}
\usage{
JAOPuTo_Core_IDA3_netposition(start, end)
}
\arguments{
\item{start}{Start datetime; (POSIXct, Date, or character convertible to POSIXct)}

\item{end}{End datetime; (POSIXct, Date, or character convertible to POSIXct)}
}
\value{
A tibble containing the combined data returned by all API requests.
}
\description{
Download Core net position for selected bidding zones in Core intraday auction 3.
}
\examples{
\dontrun{
JAOPuTo_Core_IDA3_netposition(
  start = "2025-01-01 00:00",
  end = "2025-01-10 23:00"
)
}
}
