% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_initialcomputation.R
\name{JAOPuTo_Core_initialcomputation}
\alias{JAOPuTo_Core_initialcomputation}
\title{Core - Initial Computation (Virgin Domain)}
\usage{
JAOPuTo_Core_initialcomputation(StartDateTime, EndDateTime, Presolved = TRUE)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}

\item{Presolved}{Boolean variable (TRUE/FALSE) to determine whether only presolved network elements are included (defaults to TRUE)}
}
\value{
a dataframe containing the  flow-based parameters of the selected business day and MTU of the initial flow-based computation (virgin domain, RefProg balanced).
}
\description{
Core - Initial Computation (Virgin Domain)
}
\examples{
JAOPuTo_Core_initialcomputation(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
