% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_intradayATC.R
\name{JAOPuTo_Core_intradayATC}
\alias{JAOPuTo_Core_intradayATC}
\title{Core - Intraday ATC}
\usage{
JAOPuTo_Core_intradayATC(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the Available Transfer Capacities (ATC) for the intraday timeframe on Core borders
}
\description{
Core - Intraday ATC
}
\examples{
JAOPuTo_Core_intradayATC(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
