% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_intradayNTC.R
\name{JAOPuTo_Core_intradayNTC}
\alias{JAOPuTo_Core_intradayNTC}
\title{Core - Intraday NTC}
\usage{
JAOPuTo_Core_intradayNTC(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the Net Transfer Capacities (NTC) for the intraday timeframe on Core borders
}
\description{
Core - Intraday NTC
}
\examples{
JAOPuTo_Core_intradayNTC(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
