% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_lta.R
\name{JAOPuTo_Core_longtermallocations}
\alias{JAOPuTo_Core_longtermallocations}
\title{Core - Long-Term Allocations}
\usage{
JAOPuTo_Core_longtermallocations(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the long-term allocations on Core bidding zone borders
}
\description{
Core - Long-Term Allocations
}
\examples{
JAOPuTo_Core_longtermallocations(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
