% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_finalmaxexchanges.R
\name{JAOPuTo_Core_maxexchangerestrictions}
\alias{JAOPuTo_Core_maxexchangerestrictions}
\title{Core - Maximum Final bilateral exchange restrictions}
\usage{
JAOPuTo_Core_maxexchangerestrictions(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the final bilateral exchange restrictions between each combination of Core bidding zones
}
\description{
Core - Maximum Final bilateral exchange restrictions
}
\examples{
JAOPuTo_Core_maxexchangerestrictions(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
