% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_maxexchanges.R
\name{JAOPuTo_Core_maxexchanges}
\alias{JAOPuTo_Core_maxexchanges}
\title{Core - Maximum Bilateral exchanges}
\usage{
JAOPuTo_Core_maxexchanges(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the maximum bilateral exchanges ("BEX") between each combination of Core bidding zones
}
\description{
Core - Maximum Bilateral exchanges
}
\examples{
JAOPuTo_Core_maxexchanges(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
