% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_maxnetpositions.R
\name{JAOPuTo_Core_maxnetpositions}
\alias{JAOPuTo_Core_maxnetpositions}
\title{Core - Maximum Net Positions}
\usage{
JAOPuTo_Core_maxnetpositions(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the minimum and maximum net positions for all Core bidding zones
}
\description{
Core - Maximum Net Positions
}
\examples{
JAOPuTo_Core_maxnetpositions(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
