% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_prefinalcomputation.R
\name{JAOPuTo_Core_prefinalcomputatation}
\alias{JAOPuTo_Core_prefinalcomputatation}
\title{Core - Pre-final Computation (Early Publication)}
\usage{
JAOPuTo_Core_prefinalcomputatation(
  StartDateTime,
  EndDateTime,
  Presolved = TRUE
)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}

\item{Presolved}{Boolean variable (TRUE/FALSE) to determine whether only presolved network elements are included (defaults to TRUE)}
}
\value{
a dataframe containing the pre final flow-based parameters of the selected business day and MTU before long term nominations (zero balanced).
}
\description{
Core - Pre-final Computation (Early Publication)
}
\examples{
JAOPuTo_Core_prefinalcomputatation(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
