% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_pricespreads.R
\name{JAOPuTo_Core_pricespreads}
\alias{JAOPuTo_Core_pricespreads}
\title{Core - Price Spreads}
\usage{
JAOPuTo_Core_pricespreads(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the price spreads (in €/MWh) on Core and non-Core bidding zone borders
}
\description{
Core - Price Spreads
}
\examples{
JAOPuTo_Core_pricespreads(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
