% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAOPuTo_Core_rapreventive.R
\name{JAOPuTo_Core_rapreventive}
\alias{JAOPuTo_Core_rapreventive}
\title{Core - Remedial Actions Preventive}
\usage{
JAOPuTo_Core_rapreventive(start, end)
}
\arguments{
\item{start}{Start datetime; (POSIXct, Date, or character convertible to POSIXct)}

\item{end}{End datetime; (POSIXct, Date, or character convertible to POSIXct)}
}
\value{
A tibble containing the combined data returned by all API requests.
}
\description{
Download preventive remedial actions applied in Core DA FBMC.
}
\examples{
\dontrun{
JAOPuTo_Core_rapreventive(
  start = "2025-01-01 00:00",
  end = "2025-01-10 23:00"
)
}11
}
