% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_referencenetpositions.R
\name{JAOPuTo_Core_refnetpositions}
\alias{JAOPuTo_Core_refnetpositions}
\title{Core - Reference Net Positions}
\usage{
JAOPuTo_Core_refnetpositions(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the reference net positions for non-Core bidding zones
}
\description{
Core - Reference Net Positions
}
\examples{
JAOPuTo_Core_refnetpositions(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
