% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_refprog.R
\name{JAOPuTo_Core_refprog}
\alias{JAOPuTo_Core_refprog}
\title{Core - Reference Programs}
\usage{
JAOPuTo_Core_refprog(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the exchange data per border that are used for merging of the European grid models including HVDCinterconnectors within the synchronous area.
}
\description{
Core - Reference Programs
}
\examples{
JAOPuTo_Core_refprog(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
