% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_remedialactionscurative.R
\name{JAOPuTo_Core_remedialactionscurative}
\alias{JAOPuTo_Core_remedialactionscurative}
\title{Core - Remedial Actions (Curative)}
\usage{
JAOPuTo_Core_remedialactionscurative(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the curative remedial actions applied by Core TSOs
}
\description{
Core - Remedial Actions (Curative)
}
\examples{
JAOPuTo_Core_remedialactionscurative(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
