% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_scheduledexchanges.R
\name{JAOPuTo_Core_scheduledexchanges}
\alias{JAOPuTo_Core_scheduledexchanges}
\title{Core - Scheduled exchanges}
\usage{
JAOPuTo_Core_scheduledexchanges(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the scheduled bilateral exchanges between each combination of Core bidding zones
}
\description{
Core - Scheduled exchanges
}
\examples{
JAOPuTo_Core_scheduledexchanges(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
