% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_SGM.R
\name{JAOPuTo_Core_SGM}
\alias{JAOPuTo_Core_SGM}
\title{Core - SGM}
\usage{
JAOPuTo_Core_SGM()
}
\value{
a dataframe containing the Core Static Grid Model, including lines and transformers (source: SGM 6th release)
}
\description{
Core - SGM
}
\examples{
JAOPuTo_Core_SGM()
}
