% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_shadowauctionsATC.R
\name{JAOPuTo_Core_shadowauctionsATC}
\alias{JAOPuTo_Core_shadowauctionsATC}
\title{Core - Shadow Auction ATC}
\usage{
JAOPuTo_Core_shadowauctionsATC(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the Available Transfer Capacities (ATC) for Shadow Auctions on Core borders
}
\description{
Core - Shadow Auction ATC
}
\examples{
JAOPuTo_Core_shadowauctionsATC(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
