% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_shadowprices.R
\name{JAOPuTo_Core_shadowprices}
\alias{JAOPuTo_Core_shadowprices}
\title{Core - Shadow Prices}
\usage{
JAOPuTo_Core_shadowprices(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the binding constraints (CNECs) after Market Coupling, with its shadow price.
}
\description{
Core - Shadow Prices
}
\examples{
JAOPuTo_Core_shadowprices(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
