% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_spanningDFP.R
\name{JAOPuTo_Core_spanningDFP}
\alias{JAOPuTo_Core_spanningDFP}
\title{Core - Spanning / DFP}
\usage{
JAOPuTo_Core_spanningDFP(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe reflecting the data completeness for each of the data items (parameters) that can be consulted through the Publication Tool
}
\description{
Core - Spanning / DFP
}
\examples{
JAOPuTo_Core_spanningDFP(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
