% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core_validationreductions.R
\name{JAOPuTo_Core_validationreductions}
\alias{JAOPuTo_Core_validationreductions}
\title{Core - Validation Reductions}
\usage{
JAOPuTo_Core_validationreductions(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the network elements for which capacity has been reduced as an outcome of the validation processes, including a justification for this reduction
}
\description{
Core - Validation Reductions
}
\examples{
JAOPuTo_Core_validationreductions(as.POSIXct("2024-01-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-01-01 23:00", "Europe/Brussels"))
}
