% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreID_ID3_congestionincome.R
\name{JAOPuTo_CoreID_ID3_congestionincome}
\alias{JAOPuTo_CoreID_ID3_congestionincome}
\title{Core ID - ID3 - Congestion Income}
\usage{
JAOPuTo_CoreID_ID3_congestionincome(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the gross and net congestion income per Core hub, TSO and border
}
\description{
Core ID - ID3 - Congestion Income
}
\examples{
JAOPuTo_CoreID_ID3_congestionincome(as.POSIXct("2024-07-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-07-01 23:00", "Europe/Brussels"))
}
