% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreID_ID3_pricespreads.R
\name{JAOPuTo_CoreID_ID3_pricespreads}
\alias{JAOPuTo_CoreID_ID3_pricespreads}
\title{Core - ID3 - Price Spreads}
\usage{
JAOPuTo_CoreID_ID3_pricespreads(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the price spreads (in €/MWh) on Core and non-Core bidding zone borders
}
\description{
Core - ID3 - Price Spreads
}
\examples{
JAOPuTo_CoreID_ID3_pricespreads(as.POSIXct("2024-07-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-07-01 23:00", "Europe/Brussels"))
}
