% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAOPuTo_CoreID_IDA1_scheduledexchanges.R
\name{JAOPuTo_CoreID_IDA1_scheduledexchanges}
\alias{JAOPuTo_CoreID_IDA1_scheduledexchanges}
\title{Core IDA1 - Sheduled Exchanges}
\usage{
JAOPuTo_CoreID_IDA1_scheduledexchanges(start, end)
}
\arguments{
\item{start}{Start datetime; (POSIXct, Date, or character convertible to POSIXct)}

\item{end}{End datetime; (POSIXct, Date, or character convertible to POSIXct)}
}
\value{
A tibble containing the combined data returned by all API requests.
}
\description{
Download  bilateral scheduled exchanges between selected bidding zones in Core intraday auction 1.
}
\examples{
\dontrun{
JAOPuTo_CoreID_IDA1_scheduledexchanges(
  start = "2025-01-01 00:00",
  end = "2025-01-10 23:00"
)
}
}
