% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAOPuTo_CoreID_IDA3_congestionincome.R
\name{JAOPuTo_CoreID_IDA3_congestionincome}
\alias{JAOPuTo_CoreID_IDA3_congestionincome}
\title{Corev IDA3 - Congestion Income}
\usage{
JAOPuTo_CoreID_IDA3_congestionincome(start, end)
}
\arguments{
\item{start}{Start datetime; (POSIXct, Date, or character convertible to POSIXct)}

\item{end}{End datetime; (POSIXct, Date, or character convertible to POSIXct)}
}
\value{
A tibble containing the combined data returned by all API requests.
}
\description{
Download congestion income generated in Core intraday auction 3.
}
\examples{
\dontrun{
JAOPuTo_CoreID_IDA3_congestionincome(
  start = "2025-01-01 00:00",
  end = "2025-01-10 23:00"
)
}
}
