% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAOPuTo_CoreID_IDA3_pricespreads.R
\name{JAOPuTo_CoreID_IDA3_pricespreads}
\alias{JAOPuTo_CoreID_IDA3_pricespreads}
\title{Core IDA3 - Price Spreads}
\usage{
JAOPuTo_CoreID_IDA3_pricespreads(start, end)
}
\arguments{
\item{start}{Start datetime; (POSIXct, Date, or character convertible to POSIXct)}

\item{end}{End datetime; (POSIXct, Date, or character convertible to POSIXct)}
}
\value{
A tibble containing the combined data returned by all API requests.
}
\description{
Download price spreads between selected bidding zones in Core intraday auction 3.
}
\examples{
\dontrun{
JAOPuTo_CoreID_IDA3_pricespreads(
  start = "2025-01-01 00:00",
  end = "2025-01-10 23:00"
)
}
}
