% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreID_IDCCa_initialATC.R
\name{JAOPuTo_CoreID_IDCCa_initialATC}
\alias{JAOPuTo_CoreID_IDCCa_initialATC}
\title{CoreID - IDCC(a) - Initial ATCs for SIDC}
\usage{
JAOPuTo_CoreID_IDCCa_initialATC(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the initial leftover available transmission capacity after the day-ahead market before TSO validation for defined borders in both directions in MW.
}
\description{
CoreID - IDCC(a) - Initial ATCs for SIDC
}
\examples{
JAOPuTo_CoreID_IDCCa_initialATC(as.POSIXct("2024-07-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-07-01 23:00", "Europe/Brussels"))
}
