% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreID_IDCCb_finalNTC.R
\name{JAOPuTo_CoreID_IDCCb_finalNTC}
\alias{JAOPuTo_CoreID_IDCCb_finalNTC}
\title{CoreID - IDCC(b) - Final NTCs for SIDC}
\usage{
JAOPuTo_CoreID_IDCCb_finalNTC(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the final leftover net transmission capacity after the day-ahead market after TSO validation for defined borders in both directions in MW.
}
\description{
CoreID - IDCC(b) - Final NTCs for SIDC
}
\examples{
JAOPuTo_CoreID_IDCCb_finalNTC(as.POSIXct("2024-07-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-07-01 23:00", "Europe/Brussels"))
}
