% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreID_IDCCb_maxexchanges.R
\name{JAOPuTo_CoreID_IDCCb_maxexchanges}
\alias{JAOPuTo_CoreID_IDCCb_maxexchanges}
\title{CoreID - IDCCb - Maximum Bilateral exchanges}
\usage{
JAOPuTo_CoreID_IDCCb_maxexchanges(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the maximum bilateral exchanges ("BEX") between each combination of Core bidding zones
}
\description{
CoreID - IDCCb - Maximum Bilateral exchanges
}
\examples{
JAOPuTo_CoreID_IDCCb_maxexchanges(as.POSIXct("2024-07-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-07-01 23:00", "Europe/Brussels"))
}
