% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAOPuTo_CoreID_IDCCb_maxnetpos.R
\name{JAOPuTo_CoreID_IDCCb_maxnetpos}
\alias{JAOPuTo_CoreID_IDCCb_maxnetpos}
\title{Core IDCC(b) - Max Net Positions}
\usage{
JAOPuTo_CoreID_IDCCb_maxnetpos(start, end)
}
\arguments{
\item{start}{Start datetime; (POSIXct, Date, or character convertible to POSIXct)}

\item{end}{End datetime; (POSIXct, Date, or character convertible to POSIXct)}
}
\value{
A tibble containing the combined data returned by all API requests.
}
\description{
Download maximum import and export position for selected bidding zones in Core IDCC(b).
}
\examples{
\dontrun{
JAOPuTo_CoreID_IDCCb_maxnetpos(
  start = "2025-01-01 00:00",
  end = "2025-01-10 23:00"
)
}
}
