% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreID_IDCCb_maxnetpositions.R
\name{JAOPuTo_CoreID_IDCCb_maxnetpositions}
\alias{JAOPuTo_CoreID_IDCCb_maxnetpositions}
\title{CoreID - IDCC(b) - Maximum Net Positions}
\usage{
JAOPuTo_CoreID_IDCCb_maxnetpositions(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the minimum and maximum net positions for all Core bidding zones
}
\description{
CoreID - IDCC(b) - Maximum Net Positions
}
\examples{
JAOPuTo_CoreID_IDCCb_maxnetpositions(as.POSIXct("2024-07-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-07-01 23:00", "Europe/Brussels"))
}
