% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreID_IDCCb_referencenetpositions.R
\name{JAOPuTo_CoreID_IDCCb_refnetpositions}
\alias{JAOPuTo_CoreID_IDCCb_refnetpositions}
\title{CoreID - IDCC(b) - Reference Net Positions}
\usage{
JAOPuTo_CoreID_IDCCb_refnetpositions(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the reference net positions for non-Core bidding zones
}
\description{
CoreID - IDCC(b) - Reference Net Positions
}
\examples{
JAOPuTo_CoreID_IDCCb_refnetpositions(as.POSIXct("2024-07-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-07-01 23:00", "Europe/Brussels"))
}
