% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreID_IDCCb_refprog.R
\name{JAOPuTo_CoreID_IDCCb_refprog}
\alias{JAOPuTo_CoreID_IDCCb_refprog}
\title{CoreID - IDCCb - Reference Programs}
\usage{
JAOPuTo_CoreID_IDCCb_refprog(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the exchange data per border that are used for merging of the European grid models including HVDCinterconnectors within the synchronous area.
}
\description{
CoreID - IDCCb - Reference Programs
}
\examples{
JAOPuTo_CoreID_IDCCb_refprog(as.POSIXct("2024-07-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-07-01 23:00", "Europe/Brussels"))
}
