% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreID_IDCCb_validationreductions.R
\name{JAOPuTo_CoreID_IDCCb_validationreductions}
\alias{JAOPuTo_CoreID_IDCCb_validationreductions}
\title{CoreID - IDCCb - Validation Reductions}
\usage{
JAOPuTo_CoreID_IDCCb_validationreductions(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the network elements for which capacity has been reduced as an outcome of the validation processes, including a justification for this reduction
}
\description{
CoreID - IDCCb - Validation Reductions
}
\examples{
JAOPuTo_CoreID_IDCCb_validationreductions(as.POSIXct("2024-07-01 00:00", "Europe/Brussels"),
as.POSIXct("2024-07-01 23:00", "Europe/Brussels"))
}
