% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D2CF.R
\name{JAOPuTo_D2CF}
\alias{JAOPuTo_D2CF}
\title{D2CF}
\usage{
JAOPuTo_D2CF(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the D2CF (Two Day-Ahead Congestion Forecast) for Core bidding zones
}
\description{
D2CF
}
\examples{
df <- JAOPuTo_D2CF(as.POSIXct("2022-06-09 00:00", "CET"), as.POSIXct("2022-12-31 23:00", "CET"))
}
