% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalBEXrestrictions.R
\name{JAOPuTo_finalBEXrestrictions}
\alias{JAOPuTo_finalBEXrestrictions}
\title{Final Bilateral Exchange Restrictions}
\usage{
JAOPuTo_finalBEXrestrictions(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the final bilateral exchange restrictions on borders between Core bidding zones
}
\description{
Final Bilateral Exchange Restrictions
}
\examples{
df <- JAOPuTo_finalBEXrestrictions(as.POSIXct("2022-06-09 00:00", "CET"), as.POSIXct("2022-12-31 23:00", "CET"))
}
