% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalcomputation.R
\name{JAOPuTo_finaldomain}
\alias{JAOPuTo_finaldomain}
\title{Final Computation}
\usage{
JAOPuTo_finaldomain(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the final computation flow-based domain for Core TSOs
}
\description{
Final Computation
}
\examples{
df <- JAOPuTo_finaldomain(as.POSIXct("2022-06-09 00:00", "CET"), as.POSIXct("2022-06-09 23:00", "CET"))
}
