% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intradayATC.R
\name{JAOPuTo_intradayATC}
\alias{JAOPuTo_intradayATC}
\title{Intraday ATC}
\usage{
JAOPuTo_intradayATC(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the intraday ATC values on borders between Core bidding zones
}
\description{
Intraday ATC
}
\examples{
df <- JAOPuTo_intradayATC(as.POSIXct("2022-06-09 00:00", "CET"), as.POSIXct("2022-12-31 23:00", "CET"))
}
