% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intradayNTC.R
\name{JAOPuTo_intradayNTC}
\alias{JAOPuTo_intradayNTC}
\title{Intraday NTC}
\usage{
JAOPuTo_intradayNTC(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the intraday NTC values on borders between Core bidding zones
}
\description{
Intraday NTC
}
\examples{
df <- JAOPuTo_intradayNTC(as.POSIXct("2022-06-09 00:00", "CET"), as.POSIXct("2022-12-31 23:00", "CET"))
}
