% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LTA.R
\name{JAOPuTo_LTA}
\alias{JAOPuTo_LTA}
\title{Long-Term Allocations}
\usage{
JAOPuTo_LTA(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the LTA (Long-Term Allocations) values on borders between Core bidding zones
}
\description{
Long-Term Allocations
}
\examples{
df <- JAOPuTo_LTA(as.POSIXct("2022-06-09 00:00", "CET"), as.POSIXct("2022-12-31 23:00", "CET"))
}
