% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxexchanges.R
\name{JAOPuTo_maxexchanges}
\alias{JAOPuTo_maxexchanges}
\title{Max exchanges}
\usage{
JAOPuTo_maxexchanges(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the max bilateral exchanges between Core bidding zones
}
\description{
Max exchanges
}
\examples{
df <- JAOPuTo_maxexchanges(as.POSIXct("2022-06-09 00:00", "CET"), as.POSIXct("2022-12-31 23:00", "CET"))
}
