% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxnetpositions.R
\name{JAOPuTo_maxnetpositions}
\alias{JAOPuTo_maxnetpositions}
\title{Core min-max positions}
\usage{
JAOPuTo_maxnetpositions(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the Core minimum and maximum net positions for all 14 Core bidding zones
}
\description{
Core min-max positions
}
\examples{
df <- JAOPuTo_maxnetpositions(as.POSIXct("2022-06-09 00:00", "CET"), as.POSIXct("2022-12-31 23:00", "CET"))
}
