% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netpositions.R
\name{JAOPuTo_netpositions}
\alias{JAOPuTo_netpositions}
\title{Core net positions}
\usage{
JAOPuTo_netpositions(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the Core net positions for all 14 Core bidding zones
}
\description{
Core net positions
}
\examples{
df <- JAOPuTo_netpositions(as.POSIXct("2022-06-09 00:00", "CET"), as.POSIXct("2022-12-31 23:00", "CET"))
}
