% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nordic_D2CF.R
\name{JAOPuTo_Nordic_D2CF}
\alias{JAOPuTo_Nordic_D2CF}
\title{Nordic - D2CF}
\usage{
JAOPuTo_Nordic_D2CF(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the aggregated assumptions from the grid model (D2CF)
}
\description{
Nordic - D2CF
}
\examples{
JAOPuTo_Nordic_D2CF(as.POSIXct("2024-10-30 00:00", "Europe/Brussels"),
as.POSIXct("2024-10-30 23:00", "Europe/Brussels"))
}
