% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nordic_maxborderflow.R
\name{JAOPuTo_Nordic_maxborderflow}
\alias{JAOPuTo_Nordic_maxborderflow}
\title{Nordic - Maximum Border Flow}
\usage{
JAOPuTo_Nordic_maxborderflow(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the maximum border flow (MaxBFlow) between each combination of Nordic bidding zones
}
\description{
Nordic - Maximum Border Flow
}
\examples{
JAOPuTo_Nordic_maxborderflow(as.POSIXct("2024-10-30 00:00", "Europe/Brussels"),
as.POSIXct("2024-10-30 23:00", "Europe/Brussels"))
}
