% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nordic_maxexchanges.R
\name{JAOPuTo_Nordic_maxexchanges}
\alias{JAOPuTo_Nordic_maxexchanges}
\title{Nordic - Maximum Bilateral exchanges}
\usage{
JAOPuTo_Nordic_maxexchanges(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the maximum bilateral exchanges ("BEX") between each combination of Nordic bidding zones
}
\description{
Nordic - Maximum Bilateral exchanges
}
\examples{
JAOPuTo_Nordic_maxexchanges(as.POSIXct("2024-10-30 00:00", "Europe/Brussels"),
as.POSIXct("2024-10-30 23:00", "Europe/Brussels"))
}
