% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nordic_maxnetpositions.R
\name{JAOPuTo_Nordic_maxnetpositions}
\alias{JAOPuTo_Nordic_maxnetpositions}
\title{Nordic - Nordic Net Positions}
\usage{
JAOPuTo_Nordic_maxnetpositions(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the minimum and maximum net positions for all Nordic bidding zones
}
\description{
Nordic - Nordic Net Positions
}
\examples{
JAOPuTo_Nordic_maxnetpositions(as.POSIXct("2024-10-30 00:00", "Europe/Brussels"),
as.POSIXct("2024-10-30 23:00", "Europe/Brussels"))
}
