% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nordic_shadowprices.R
\name{JAOPuTo_Nordic_shadowprices}
\alias{JAOPuTo_Nordic_shadowprices}
\title{Nordic - Shadow Prices}
\usage{
JAOPuTo_Nordic_shadowprices(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the binding constraints (CNECs) after Market Coupling, with its shadow price.
}
\description{
Nordic - Shadow Prices
}
\examples{
JAOPuTo_Nordic_shadowprices(as.POSIXct("2024-10-30 00:00", "Europe/Brussels"),
as.POSIXct("2024-10-30 23:00", "Europe/Brussels"))
}
