% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nordic_spanningDFP.R
\name{JAOPuTo_Nordic_spanningDFP}
\alias{JAOPuTo_Nordic_spanningDFP}
\title{Nordic - Spanning / DFP}
\usage{
JAOPuTo_Nordic_spanningDFP(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe reflecting the data completeness for each of the data items (parameters) that can be consulted through the Publication Tool
}
\description{
Nordic - Spanning / DFP
}
\examples{
JAOPuTo_Nordic_spanningDFP(as.POSIXct("2024-10-30 00:00", "Europe/Brussels"),
as.POSIXct("2024-10-30 23:00", "Europe/Brussels"))
}
