% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nordic_validationreductions.R
\name{JAOPuTo_Nordic_validationreductions}
\alias{JAOPuTo_Nordic_validationreductions}
\title{Nordic - Validation Reductions}
\usage{
JAOPuTo_Nordic_validationreductions(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the network elements for which capacity has been reduced as an outcome of the validation processes, including a justification for this reduction
}
\description{
Nordic - Validation Reductions
}
\examples{
JAOPuTo_Nordic_validationreductions(as.POSIXct("2024-11-05 00:00", "Europe/Brussels"),
as.POSIXct("2024-11-05 23:00", "Europe/Brussels"))
}
