% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadowprices.R
\name{JAOPuTo_shadowprices}
\alias{JAOPuTo_shadowprices}
\title{Shadow Prices}
\usage{
JAOPuTo_shadowprices(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the shadow prices per active constraint for Core TSOs
}
\description{
Shadow Prices
}
\examples{
df <- JAOPuTo_shadowprices(as.POSIXct("2022-06-09 00:00", "CET"), as.POSIXct("2022-12-31 23:00", "CET"))
}
