% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationreductions.R
\name{JAOPuTo_validationreductions}
\alias{JAOPuTo_validationreductions}
\title{Validation Reductions}
\usage{
JAOPuTo_validationreductions(StartDateTime, EndDateTime)
}
\arguments{
\item{StartDateTime}{POSIXct-class variable containing the start datetime for data download}

\item{EndDateTime}{POSIXct-class variable containing the end datetime for data download}
}
\value{
a dataframe containing the validation reductions on presolved CNECs
}
\description{
Validation Reductions
}
\examples{
df <- JAOPuTo_validationreductions(as.POSIXct("2022-06-09 00:00", "CET"), as.POSIXct("2022-12-31 23:00", "CET"))
}
