% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAOPuTo_Core_D2CF.R
\name{JAOPuTo_Core_D2CF}
\alias{JAOPuTo_Core_D2CF}
\title{Core - Two-Day Ahead Congestion Forecast (D2CF)}
\usage{
JAOPuTo_Core_D2CF(start, end)
}
\arguments{
\item{start}{Start datetime; (POSIXct, Date, or character convertible to POSIXct)}

\item{end}{End datetime; (POSIXct, Date, or character convertible to POSIXct)}
}
\value{
A tibble containing the combined data returned by all API requests.
}
\description{
Download D2CF for Core bidding zone borders in Core DA FBMC.
}
\examples{
\dontrun{
JAOPuTo_Core_D2CF(
  start = "2025-01-01 00:00",
  end = "2025-01-10 23:00"
)
}
}
