% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAOPuTo_Core_finalcomputation.R
\name{JAOPuTo_Core_finalcomputation}
\alias{JAOPuTo_Core_finalcomputation}
\title{Core - Final Computation}
\usage{
JAOPuTo_Core_finalcomputation(start, end, presolved = TRUE)
}
\arguments{
\item{start}{Start datetime; (POSIXct, Date, or character convertible to POSIXct)}

\item{end}{End datetime; (POSIXct, Date, or character convertible to POSIXct)}

\item{presolved}{Boolean; filter only presolved CNECs (default = TRUE)}
}
\value{
A tibble containing the combined data returned by all API requests.
}
\description{
Download final flow-based parameters in Core DA FBMC.
}
\examples{
\dontrun{
JAOPuTo_Core_finalcomputation(
  start = "2025-01-01 00:00",
  end = "2025-01-01 01:00"
)
}
}
