% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAOPuTo_initialcomputation.R
\name{JAOPuTo_Core_initialcomputation}
\alias{JAOPuTo_Core_initialcomputation}
\title{Core - Initial Computation Computation (Virgin Domain)}
\usage{
JAOPuTo_Core_initialcomputation(start, end, presolved = TRUE)
}
\arguments{
\item{start}{Start datetime; (POSIXct, Date, or character convertible to POSIXct)}

\item{end}{End datetime; (POSIXct, Date, or character convertible to POSIXct)}

\item{presolved}{Boolean; filter only presolved CNECs (default = TRUE)}
}
\value{
A tibble containing the combined data returned by all API requests.
}
\description{
Download pre-final flow-based parameters (early publication) in Core DA FBMC.
}
\examples{
\dontrun{
JAOPuTo_Core_initialcomputation(
  start = "2025-01-01 00:00",
  end = "2025-01-01 01:00"
)
}
}
