% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAOPuTo_Core_intradayatc.R
\name{JAOPuTo_Core_intradayatc}
\alias{JAOPuTo_Core_intradayatc}
\title{Core - Intraday ATC}
\usage{
JAOPuTo_Core_intradayatc(start, end)
}
\arguments{
\item{start}{Start datetime; (POSIXct, Date, or character convertible to POSIXct)}

\item{end}{End datetime; (POSIXct, Date, or character convertible to POSIXct)}
}
\value{
A tibble containing the combined data returned by all API requests.
}
\description{
Download Intraday ATCs on bidding zone borders in Core DA FBMC.
}
\examples{
\dontrun{
JAOPuTo_Core_intradayatc(
  start = "2024-01-01 00:00",
  end = "2024-01-10 23:00"
)
}
}
