% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAOPuTo_Core_intradayntc.R
\name{JAOPuTo_Core_intradayntc}
\alias{JAOPuTo_Core_intradayntc}
\title{Core - Intraday NTC}
\usage{
JAOPuTo_Core_intradayntc(start, end)
}
\arguments{
\item{start}{Start datetime; (POSIXct, Date, or character convertible to POSIXct)}

\item{end}{End datetime; (POSIXct, Date, or character convertible to POSIXct)}
}
\value{
A tibble containing the combined data returned by all API requests.
}
\description{
Download Intraday NTCs on bidding zone borders in Core DA FBMC.
}
\examples{
\dontrun{
JAOPuTo_Core_intradayntc(
  start = "2024-01-01 00:00",
  end = "2024-01-10 23:00"
)
}
}
