% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAOPuTo_Core_refprog.R
\name{JAOPuTo_Core_refprog}
\alias{JAOPuTo_Core_refprog}
\title{Core - Reference Program}
\usage{
JAOPuTo_Core_refprog(start, end)
}
\arguments{
\item{start}{Start datetime; (POSIXct, Date, or character convertible to POSIXct)}

\item{end}{End datetime; (POSIXct, Date, or character convertible to POSIXct)}
}
\value{
A tibble containing the combined data returned by all API requests.
}
\description{
Download reference programs of Core and non-Core bidding zone borders in Core DA FBMC.
}
\examples{
\dontrun{
JAOPuTo_Core_refprog(
  start = "2025-01-01 00:00",
  end = "2025-01-10 23:00"
)
}
}
