% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAOPuTo_Core_spanningdfp.R
\name{JAOPuTo_Core_spanningdfp}
\alias{JAOPuTo_Core_spanningdfp}
\title{Core - Spanning / Default Flow-based Parameters}
\usage{
JAOPuTo_Core_spanningdfp(start, end)
}
\arguments{
\item{start}{Start datetime; (POSIXct, Date, or character convertible to POSIXct)}

\item{end}{End datetime; (POSIXct, Date, or character convertible to POSIXct)}
}
\value{
A tibble containing the combined data returned by all API requests.
}
\description{
Download status of spanning or default parameter application in Core DA FBMC.
}
\examples{
\dontrun{
JAOPuTo_Core_spanningdfp(
  start = "2025-06-25 00:00",
  end = "2025-06-25 23:00"
)
}
}
