% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAOPuTo_CoreID_IDCCb_maxexchanges.R
\name{JAOPuTo_CoreID_IDCCb_maxexchanges}
\alias{JAOPuTo_CoreID_IDCCb_maxexchanges}
\title{Core IDCC(b) - Max Exchanges (MaxBex)}
\usage{
JAOPuTo_CoreID_IDCCb_maxexchanges(start, end)
}
\arguments{
\item{start}{Start datetime; (POSIXct, Date, or character convertible to POSIXct)}

\item{end}{End datetime; (POSIXct, Date, or character convertible to POSIXct)}
}
\value{
A tibble containing the combined data returned by all API requests.
}
\description{
Download maximum bilateral exchanges between selected bidding zones in Core IDCC(b).
}
\examples{
\dontrun{
JAOPuTo_CoreID_IDCCb_maxexchanges(
  start = "2025-01-01 00:00",
  end = "2025-01-10 23:00"
)
}
}
