% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domainvisualization_position.R
\name{JAOPuTo_domainvisualization_positions}
\alias{JAOPuTo_domainvisualization_positions}
\title{Visualizing the flow-based domains}
\usage{
JAOPuTo_domainvisualization_positions(
  DateTime,
  BiddingZone1,
  BiddingZone2,
  Reference = "MCP",
  ReturnDataframes = FALSE
)
}
\arguments{
\item{DateTime}{POSIXct-class variable containing the datetime for which to visualize the flow-based domains}

\item{BiddingZone1}{character class variable determining the first Core bidding zone for which to visualize the possible net positions (horizontally)}

\item{BiddingZone2}{character class variable determining the second Core bidding zone for which to visualize the possible net positions (vertically)}

\item{Reference}{character-class variable (default: "MCP", alternative: "Zero-balanced") determining which Core net positions to assume for the cross-section of the flow-based domain}

\item{ReturnDataframes}{boolean variable (default: FALSE) determining whether to return the dataframes}
}
\value{
a ggplot2 object and, if ReturnDataframes == TRUE, the dataframes to build the flow-based domain
}
\description{
Visualizing the flow-based domains
}
\examples{
JAOPuTo_domainvisualization_positions(as.POSIXct("2022-06-09 00:00", "CET"), "Belgium", "France", "MCP", TRUE)
}
